third_party/blink/web_tests/external/wpt/css/css-inline/parsing/baseline-shift-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['baseline-shift'] = "5" should not set the property value
  assert_equals: expected "" but got "5"
Harness: the test ran to completion.
